+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Ribbon V2.11.0, RibbonTools V1.3.6
+
+#### Changed (RibbonTools)
+
+- Bugfix: GalleryControls in AppMenuGroup and in other Groups
 ### Ribbon V2.11.0, RibbonTools V1.3.5
 
 #### Changed (Ribbon)